+++ src/types/generated.ts
@@ -322,6 +322,7 @@ export type AwsApiGatewayRestApi = AwsBaseService & {
   endpointConfiguration?: Maybe<AwsApiGatewayEndpointConfiguration>;
   minimumCompressionSize?: Maybe<Scalars['Int']>;
   policy?: Maybe<AwsIamJsonPolicy>;
+  rawPolicy?: Maybe<Scalars['String']>;
   resources?: Maybe<Array<Maybe<AwsApiGatewayResource>>>;
   route53Record?: Maybe<Array<Maybe<AwsRoute53Record>>>;
   stages?: Maybe<Array<Maybe<AwsApiGatewayStage>>>;
@@ -551,9 +552,7 @@ export type AwsBaseService = {
   region?: Maybe<Scalars['String']>;
 };
 
-export type AwsBilling = {
-  accountId: Scalars['String'];
-  id: Scalars['String'];
+export type AwsBilling = AwsOptionalService & {
   last30Days?: Maybe<Array<Maybe<AwsServiceBillingInfo>>>;
   last30DaysDailyAverage?: Maybe<Array<Maybe<AwsServiceBillingInfo>>>;
   monthToDate?: Maybe<Array<Maybe<AwsServiceBillingInfo>>>;
@@ -742,6 +741,7 @@ export type AwsCloudFormationStackSetParameter = {
 
 export type AwsCloudfront = AwsBaseService & {
   callerReference?: Maybe<Scalars['String']>;
+  cloudwatch?: Maybe<Array<Maybe<AwsCloudwatch>>>;
   customErrorResponses?: Maybe<Array<Maybe<AwsCloudfrontCustomErrorResponse>>>;
   defaultCacheBehavior?: Maybe<AwsCloudfrontCacheBehavior>;
   defaultRootObject?: Maybe<Scalars['String']>;
@@ -749,7 +749,7 @@ export type AwsCloudfront = AwsBaseService & {
   elb?: Maybe<Array<Maybe<AwsElb>>>;
   enabled?: Maybe<Scalars['String']>;
   etag?: Maybe<Scalars['String']>;
-  geoRestrictions?: Maybe<Scalars['String']>;
+  geoRestriction?: Maybe<AwsCloudfrontGeoRestriction>;
   httpVersion?: Maybe<Scalars['String']>;
   ipv6Enabled?: Maybe<Scalars['String']>;
   lastModified?: Maybe<Scalars['String']>;
@@ -795,6 +795,11 @@ export type AwsCloudfrontCustomOriginConfig = {
   originProtocolPolicy?: Maybe<Scalars['String']>;
   originReadTimeout?: Maybe<Scalars['Int']>;
   originSslProtocols?: Maybe<AwsCloudfrontOriginSslProtocols>;
+};
+
+export type AwsCloudfrontGeoRestriction = {
+  locations?: Maybe<Array<Maybe<Scalars['String']>>>;
+  restrictionType?: Maybe<Scalars['String']>;
 };
 
 export type AwsCloudfrontLoggingConfig = {
@@ -892,6 +897,7 @@ export type AwsCloudtrailStatus = {
 export type AwsCloudwatch = AwsBaseService & {
   actions?: Maybe<Array<Maybe<Scalars['String']>>>;
   actionsEnabled?: Maybe<Scalars['String']>;
+  cloudfront?: Maybe<Array<Maybe<AwsCloudfront>>>;
   cloudtrail?: Maybe<Array<Maybe<AwsCloudtrail>>>;
   cloudwatchLog?: Maybe<Array<Maybe<AwsCloudwatchLog>>>;
   comparisonOperator?: Maybe<Scalars['String']>;
@@ -2632,6 +2638,7 @@ export type AwsElasticSearchDomain = AwsBaseService & {
   logPublishingOptions?: Maybe<Array<Maybe<AwsElasticSearchLogPublishingOption>>>;
   nodeToNodeEncryptionOptions?: Maybe<AwsElasticSearchNodeToNodeEncryptionOptions>;
   processing?: Maybe<Scalars['Boolean']>;
+  rawPolicy?: Maybe<Scalars['String']>;
   securityGroups?: Maybe<Array<Maybe<AwsSecurityGroup>>>;
   serviceSoftwareOptions?: Maybe<AwsElasticSearchServiceSoftwareOptions>;
   snapshotOptions?: Maybe<AwsElasticSearchSnapshotOptions>;
@@ -2863,7 +2870,7 @@ export type AwsFederatedAuthentication = {
   selfServiceSamlProviderArn?: Maybe<Scalars['String']>;
 };
 
-export type AwsFlowLog = AwsOptionalService & {
+export type AwsFlowLog = AwsBaseService & {
   creationTime?: Maybe<Scalars['String']>;
   deliverLogsErrorMessage?: Maybe<Scalars['String']>;
   deliverLogsPermissionArn?: Maybe<Scalars['String']>;
@@ -3071,6 +3078,7 @@ export type AwsIamPolicy = AwsBaseService & {
   name?: Maybe<Scalars['String']>;
   path?: Maybe<Scalars['String']>;
   policyContent?: Maybe<AwsIamJsonPolicy>;
+  rawPolicy?: Maybe<Scalars['String']>;
   tags?: Maybe<Array<Maybe<AwsRawTag>>>;
 };
 
@@ -3108,6 +3116,7 @@ export type AwsIamRole = AwsBaseService & {
   maxSessionDuration?: Maybe<Scalars['Int']>;
   name?: Maybe<Scalars['String']>;
   path?: Maybe<Scalars['String']>;
+  rawPolicy?: Maybe<Scalars['String']>;
   rdsCluster?: Maybe<Array<Maybe<AwsRdsCluster>>>;
   rdsDbInstance?: Maybe<Array<Maybe<AwsRdsDbInstance>>>;
   s3?: Maybe<Array<Maybe<AwsS3>>>;
@@ -3243,6 +3252,7 @@ export type AwsKms = AwsBaseService & {
   managedAirflows?: Maybe<Array<Maybe<AwsManagedAirflow>>>;
   origin?: Maybe<Scalars['String']>;
   policy?: Maybe<AwsIamJsonPolicy>;
+  rawPolicy?: Maybe<Scalars['String']>;
   rdsCluster?: Maybe<Array<Maybe<AwsRdsCluster>>>;
   rdsClusterSnapshots?: Maybe<Array<Maybe<AwsRdsClusterSnapshot>>>;
   rdsDbInstance?: Maybe<Array<Maybe<AwsRdsDbInstance>>>;
@@ -3277,6 +3287,7 @@ export type AwsLambda = AwsBaseService & {
   memorySize?: Maybe<Scalars['Int']>;
   policy?: Maybe<AwsIamJsonPolicy>;
   policyRevisionId?: Maybe<Scalars['String']>;
+  rawPolicy?: Maybe<Scalars['String']>;
   reservedConcurrentExecutions?: Maybe<Scalars['Int']>;
   runtime?: Maybe<Scalars['String']>;
   s3?: Maybe<Array<Maybe<AwsS3>>>;
@@ -3752,7 +3763,6 @@ export type AwsS3 = AwsBaseService & {
   blockPublicAcls?: Maybe<Scalars['String']>;
   blockPublicPolicy?: Maybe<Scalars['String']>;
   bucketOwnerName?: Maybe<Scalars['String']>;
-  bucketPolicies?: Maybe<Array<Maybe<AwsBucketPolicy>>>;
   cloudfrontDistributions?: Maybe<Array<Maybe<AwsCloudfront>>>;
   cloudtrails?: Maybe<Array<Maybe<AwsCloudtrail>>>;
   corsConfiguration?: Maybe<Scalars['String']>;
@@ -3768,6 +3778,8 @@ export type AwsS3 = AwsBaseService & {
   managedAirflows?: Maybe<Array<Maybe<AwsManagedAirflow>>>;
   mfa?: Maybe<Scalars['String']>;
   notificationConfiguration?: Maybe<AwsS3NotificationConfiguration>;
+  policy?: Maybe<AwsIamJsonPolicy>;
+  rawPolicy?: Maybe<Scalars['String']>;
   requesterPays?: Maybe<Scalars['String']>;
   restrictPublicBuckets?: Maybe<Scalars['String']>;
   size?: Maybe<Scalars['String']>;
@@ -3984,6 +3996,7 @@ export type AwsSns = AwsBaseService & {
   displayName?: Maybe<Scalars['String']>;
   kms?: Maybe<Array<Maybe<AwsKms>>>;
   policy?: Maybe<AwsIamJsonPolicy>;
+  rawPolicy?: Maybe<Scalars['String']>;
   s3?: Maybe<Array<Maybe<AwsS3>>>;
   subscriptions?: Maybe<Array<Maybe<AwsSnsSubscription>>>;
   tags?: Maybe<Array<Maybe<AwsRawTag>>>;
@@ -4013,6 +4026,7 @@ export type AwsSqs = AwsBaseService & {
   policy?: Maybe<AwsIamJsonPolicy>;
   queueType?: Maybe<Scalars['String']>;
   queueUrl?: Maybe<Scalars['String']>;
+  rawPolicy?: Maybe<Scalars['String']>;
   receiveMessageWaitTimeSeconds?: Maybe<Scalars['String']>;
   s3?: Maybe<Array<Maybe<AwsS3>>>;
   sqsManagedSseEnabled?: Maybe<Scalars['Boolean']>;
@@ -4212,8 +4226,12 @@ export type AwsTransitGateway = AwsBaseService & {
   vpnEcmpSupport?: Maybe<Scalars['String']>;
 };
 
-export type AwsTransitGatewayAttachment = AwsBaseService & {
+export type AwsTransitGatewayAttachment = {
+  accountId: Scalars['String'];
+  arn: Scalars['String'];
   creationTime?: Maybe<Scalars['String']>;
+  id: Scalars['String'];
+  region?: Maybe<Scalars['String']>;
   resourceId?: Maybe<Scalars['String']>;
   resourceOwnerId?: Maybe<Scalars['String']>;
   resourceType?: Maybe<Scalars['String']>;